web/requirements.txt

flask==3.0.2
flask_socketio==5.3.6
eventlet==0.36.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
